try to seperate ScorePauseScene from GameScene
