+++ frontend/src/Components/Game/GameScene.js
@@ -1,22 +1,14 @@
 // eslint-disable-next-line max-classes-per-file
 import Phaser from 'phaser';
-// eslint-disable-next-line import/no-extraneous-dependencies
 import simplify from 'simplify-js';
 import dudeAsset from '../../assets/santa.png';
-import CoinLabel from './CoinLabel';
-import coinAsset from '../../assets/coin.png';
-import coinHudAsset from '../../assets/hudcoin.png';
-import pauseButton from '../../assets/pauseButton.png';
 import Settings from '../../utils/settings';
-import MeterLabel from './MeterLabel';
 import dudeAssetJSON from '../../assets/santa.json';
 import pineSaplingAsset from '../../assets/winterTheme/pineSapling.png';
 
 const GROUND_KEY = 'groundLabel';
 const COIN_KEY = 'coin';
 const OBSTACLE_KEY = 'obstacleLabel';
-const HUD_COIN_KEY = 'hudcoin';
-const PAUSE_BUTTON = 'pause';
 const PINE_SAPLING = 'pine';
 
 const gameOptions = {
@@ -44,14 +36,11 @@ class GameScene extends Phaser.Scene {
     this.coinLabel = undefined;
     this.coins = [];
     this.gameOver = false;
-    this.scorePauseScene = undefined;
     this.caracterSpeed= undefined;
   }
 
   init() {
     this.cursors = this.input.keyboard.createCursorKeys();
-    // eslint-disable-next-line no-use-before-define
-    this.scorePauseScene = this.scene.add('pause-score', ScorePauseScene, true);
   }
 
   preload() {
@@ -109,7 +98,7 @@ class GameScene extends Phaser.Scene {
              setInterval(() => {
                this.caracterSpeed += Math.log(2) / 1000;
              }, 2000);
-             this.scorePauseScene.pauseButton.on('pointerdown', () => {
+             ScorePauseScene.pauseButton.on('pointerdown', () => {
                this.scene.run('pause-menu');
              });
            }
@@ -288,10 +277,8 @@ class GameScene extends Phaser.Scene {
     santa1.play('player-slide', true);
     const spaceJustPressed = Phaser.Input.Keyboard.JustDown(this.cursors.space);
 
-    const scorePauseScene = this.scene.get('pause-score');
-
     if (localStorage.getItem('resume')) {
-      scorePauseScene.meterLabel.resumeMeter();
+      ScorePauseScene.meterLabel.resumeMeter();
       localStorage.removeItem('resume');
     }
 
@@ -409,16 +396,16 @@ class GameScene extends Phaser.Scene {
   }
 
    hitObstacle(player) {
-    this.scorePauseScene.meterLabel.pauseMeter();
+    ScorePauseScene.meterLabel.pauseMeter();
     this.scene.stop();
-    this.scorePauseScene.meterLabel.setText(
+    ScorePauseScene.meterLabel.setText(
       `GAME OVER :  \nYour Score is ${this.formatDistance(
-        this.scorePauseScene.meterLabel.timeElapsed,
+        ScorePauseScene.meterLabel.timeElapsed,
       )}`,
     );
 
     if (localStorage.getItem('token')) {
-      this.updateScore(this.scorePauseScene.meterLabel.timeElapsed);
+      this.updateScore(ScorePauseScene.meterLabel.timeElapsed);
     }
 
     this.matter.pause();
@@ -427,7 +414,7 @@ class GameScene extends Phaser.Scene {
     this.scene.stop('game-scene');
     this.scene.stop('pause-score');
     this.scene.run('game-over', {
-      score: this.formatDistance(this.scorePauseScene.meterLabel.timeElapsed),
+      score: this.formatDistance(ScorePauseScene.meterLabel.timeElapsed),
     });
   }
 
@@ -469,12 +456,12 @@ class GameScene extends Phaser.Scene {
   checkCollision(a, b) {
     if (a.label === COIN_KEY && a.gameObject !== null && a.gameObject !== undefined) {
       a.gameObject.destroy();
-      this.scorePauseScene.coinLabel.add(gameOptions.amountCoin);
+      ScorePauseScene.coinLabel.add(gameOptions.amountCoin);
     }
 
     if (b.label === COIN_KEY && b.gameObject !== null && b.gameObject !== undefined) {
       b.gameObject.destroy();
-      this.scorePauseScene.coinLabel.add(gameOptions.amountCoin);
+      ScorePauseScene.coinLabel.add(gameOptions.amountCoin);
     }
     if (a.label === OBSTACLE_KEY && a.gameObject !== null && a.gameObject !== undefined) {
       this.hitObstacle(this.santa);
@@ -486,59 +473,4 @@ class GameScene extends Phaser.Scene {
   }
 }
 
-class ScorePauseScene extends Phaser.Scene {
-  constructor() {
-    super('pause-score');
-    this.meterLabel = undefined;
-    this.pauseButton = undefined;
-    this.coinLabel = undefined;
-  }
-
-  preload() {
-    this.load.image(PAUSE_BUTTON, pauseButton);
-
-    // coins
-    this.load.image(HUD_COIN_KEY, coinHudAsset);
-    this.load.image(COIN_KEY, coinAsset);
-  }
-
-  create() {
-    this.meterLabel = this.createMeterLabel(20, 20);
-    this.meterLabel.setColor('#ffffff');
-
-    // pause btn
-    this.pauseButton = this.add.image(this.scale.width - 75, 50, PAUSE_BUTTON);
-    this.pauseButton.setInteractive({ useHandCursor: true });
-    this.pauseButton.setScale(0.8);
-
-    this.pauseButton.on('pointerdown', () => {
-      this.pauseGame();
-    });
-
-    // coins
-    this.add.image(30, 88, HUD_COIN_KEY);
-    this.coinLabel = this.createCoinLabel(45, 70);
-    this.add.existing(this.coinLabel);
-  }
-
-  createMeterLabel(x, y) {
-    const label = new MeterLabel(this, x, y);
-    this.add.existing(label);
-
-    return label;
-  }
-
-  pauseGame() {
-    this.meterLabel.pauseMeter();
-    this.scene.pause('pause-score');
-    this.scene.pause('game-scene');
-    this.scene.run('pause-menu');
-  }
-
-  createCoinLabel(x, y) {
-    const style = { fontSize: '32px', fill: '#ffffff', fontFamily: 'Arial, sans-serif' };
-    const label = new CoinLabel(this, x, y, 0, style);
-    return label;
-  }
-}
 export default GameScene;
