+++ frontend/src/Components/Game/ScoreLabel.js
@@ -6,6 +6,7 @@ const formatScore = (score) => `Score: ${score}`;
 export default class ScoreLabel extends Phaser.GameObjects.Text {
                  constructor(scene, x, y, score, style) {
                    super(scene, x, y, formatScore(score),style);
+                   // eslint-disable-next-line no-console
                    console.log('inside class', this.text);
                    this.score = score;
                  }
